git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@314 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Audio_Systems/OpenAL/as_basic.cpp
@@ -16,7 +16,7 @@ namespace enigma
   
   sound *sounds;
   size_t numSounds;
-  extern int sound_idmax;
+  extern size_t sound_idmax;
   
   static void eos_callback(void *soundID, ALuint src)
   {
